Require explicit tagging to plaintext HTTP registries
Rather than assuming a registry may either be HTTP or HTTPS,
it will now be required to explicitly tag an image to an HTTP url.

This will make it clear to users if they're downloading images
from an untrusted source. The absence of a URL scheme prefix
will assume HTTPS communications.

Signed-off-by: Eric Windisch <eric@windisch.us>