Load contrib/apparmor during integration tests
Changes to the AppArmor policy should be tested.

Note that this is somewhat dirty because we
cannot restore the host to its original configuration.

However, it should be noted that prior to this patch
series, the Docker daemon itself was loading apparmor
policy from within the tests, so this is no dirtier or
uglier than what has already been occurring. That is,
this patch maintains the status-quo. The only reason
this hasn't broken our CI servers is because we don't
make frequent changes to the AppArmor policies.

Long-term, we could load an ephemeral policy of a
unique name and specify this for 'docker run' during
tests, or extend the engine to allow for overriding
the name of the default policy.

Signed-off-by: Eric Windisch <eric@windisch.us>